requirements.txt
asgiref==3.7.2
Django==4.2
django-autoslug==1.9.9
django-cleanup==8.0.0
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-markdown==0.8.4
django-taggit==5.0.1
djangorestframework==3.14.0
Markdown==3.5.2
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
